+++ ng_trajectory/penalizers/curvature/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3.6
+"""Curvature penalizer.
+
+This penalizer finds all points that exceed the maximum
+admitted curvature.
+"""
+from .main import init, penalize, INVALID_POINTS